Add endpoint to fetch order details by ID (#147)
* Add endpoint to fetch order details by ID

Introduce a new `getOrderDetails` method in `OrderService` and its implementation in `OrderServiceImpl` to retrieve order details by ID. Add a new `OrderNotFoundException` to handle cases when an order ID is not found. Update `OrderController` to include a mapping for this endpoint.

* Add order handling functionality with multi-datasource support

Introduced order management by implementing entities, repositories, services, and mappers for orders and meals. Added multi-datasource configuration to support both client orders and company data. Exposed an API endpoint to fetch all orders for a given company.

---------

Co-authored-by: Dawid Dobosz <doboszdav@gmail.com>